NYC Building Complaints

A subset of the complaints received by the Department of Buildings (DOB) in New York City, USA.

Usage

data_building_complaints(...)

Source

https://data.cityofnewyork.us/Housing-Development/DOB-Complaints-Received/eabe-havv/about_data

Arguments

...

Arguments passed to pins::pin_read().

Value

tibble

Details

A data frame with 4,234 rows and 11 columns:

days_to_disposition

Days to disposition of the complaint

status

Status of the complaint

year_entered

Year the complaint was entered

latitude, longitude

Geographic coordinates

borough

Borough

special_district

Special district

unit

Unit dispositioning the complaint

community_board

Community board. 3-digit identifier: Borough code = first position, last 2 = community board

complaint_category

Complaint category

complaint_priority

Complaint priority

tibble print

data_building_complaints()
#> # A tibble: 4,234 x 11
#>    days_to_disposition status year_entered latitude longitude borough 
#>                  <dbl> <chr>  <fct>           <dbl>     <dbl> <fct>   
#>  1                  72 ACTIVE 2023             40.7     -74.0 Brooklyn
#>  2                   1 ACTIVE 2023             40.6     -74.0 Brooklyn
#>  3                  41 ACTIVE 2023             40.7     -73.9 Queens  
#>  4                  45 ACTIVE 2023             40.7     -73.8 Queens  
#>  5                  16 ACTIVE 2023             40.6     -74.0 Brooklyn
#>  6                  62 ACTIVE 2023             40.7     -73.8 Queens  
#>  7                  56 ACTIVE 2023             40.7     -74.0 Brooklyn
#>  8                  11 ACTIVE 2023             40.7     -74.0 Brooklyn
#>  9                  35 ACTIVE 2023             40.7     -73.8 Queens  
#> 10                  38 ACTIVE 2023             40.7     -73.9 Queens  
#> # i 4,224 more rows
#> # i 5 more variables: special_district <fct>, unit <fct>,
#> #   community_board <fct>, complaint_category <fct>, complaint_priority <fct>

glimpse()

tibble::glimpse(data_building_complaints())
#> Rows: 4,234
#> Columns: 11
#> $ days_to_disposition <dbl> 72, 1, 41, 45, 16, 62, 56, 11, 35, 38, 39, 106, 1,~
#> $ status              <chr> "ACTIVE", "ACTIVE", "ACTIVE", "ACTIVE", "ACTIVE", ~
#> $ year_entered        <fct> 2023, 2023, 2023, 2023, 2023, 2023, 2023, 2023, 20~
#> $ latitude            <dbl> 40.66173, 40.57668, 40.73242, 40.68245, 40.63156, ~
#> $ longitude           <dbl> -73.98297, -74.00453, -73.87630, -73.79367, -73.99~
#> $ borough             <fct> Brooklyn, Brooklyn, Queens, Queens, Brooklyn, Quee~
#> $ special_district    <fct> None, None, None, None, None, None, None, None, No~
#> $ unit                <fct> Q-L, Q-L, SPOPS, Q-L, BKLYN, Q-L, Q-L, SPOPS, Q-L,~
#> $ community_board     <fct> 307, 313, 404, 412, 312, 406, 306, 306, 409, 404, ~
#> $ complaint_category  <fct> 45, 45, 49, 45, 31, 45, 45, 49, 45, 45, 45, 4A, 31~
#> $ complaint_priority  <fct> B, B, C, B, C, B, B, C, B, B, B, B, C, C, B, B, B,~

Examples

# \donttest{
data_building_complaints()
#> # A tibble: 4,234 × 11
#>    days_to_disposition status year_entered latitude longitude borough 
#>                  <dbl> <chr>  <fct>           <dbl>     <dbl> <fct>   
#>  1                  72 ACTIVE 2023             40.7     -74.0 Brooklyn
#>  2                   1 ACTIVE 2023             40.6     -74.0 Brooklyn
#>  3                  41 ACTIVE 2023             40.7     -73.9 Queens  
#>  4                  45 ACTIVE 2023             40.7     -73.8 Queens  
#>  5                  16 ACTIVE 2023             40.6     -74.0 Brooklyn
#>  6                  62 ACTIVE 2023             40.7     -73.8 Queens  
#>  7                  56 ACTIVE 2023             40.7     -74.0 Brooklyn
#>  8                  11 ACTIVE 2023             40.7     -74.0 Brooklyn
#>  9                  35 ACTIVE 2023             40.7     -73.8 Queens  
#> 10                  38 ACTIVE 2023             40.7     -73.9 Queens  
#> # ℹ 4,224 more rows
#> # ℹ 5 more variables: special_district <fct>, unit <fct>,
#> #   community_board <fct>, complaint_category <fct>,
#> #   complaint_priority <fct>
# }