Daily Hotel Rate Data

A data set to predict the average daily rate for a hotel in Lisbon Portugal.

Usage

data_hotel_rates(...)

Source

https://github.com/rfordatascience/tidytuesday/tree/master/data/2020/2020-02-11

Arguments

...

Arguments passed to pins::pin_read().

Value

A tibble.

Details

Data are originally described in Antonio, de Almeida, and Nunes (2019). This version of the data is filtered for one hotel (the "Resort Hotel") and is intended as regression data set for predicting the average daily rate for a room. The data are post-2016; the 2016 data were used to have a predictor for the historical daily rates. See the hotel_rates.R file in the data-raw directory of the package to understand other filters used when creating this version of the data.

The agent and company fields were changed from random characters to use a set of random names.

The outcome column is avg_price_per_room.

License

No license was given for the data; See the reference below for source.

References

Antonio, N., de Almeida, A., and Nunes, L. (2019). Hotel booking demand datasets. Data in Brief, 22, 41-49.

Examples

# \donttest{
data_hotel_rates()
#> # A tibble: 15,402 × 30
#>    avg_price_per_room lead_time arrival_date_day_of_month
#>                 <dbl>     <dbl>                     <dbl>
#>  1              110         241                         2
#>  2               74         273                         2
#>  3               81.9       248                         2
#>  4               81         236                         2
#>  5              112.        243                         2
#>  6               90.8       267                         2
#>  7              317          94                         2
#>  8              159          10                         2
#>  9              184         156                         2
#> 10              107.        170                         2
#> # ℹ 15,392 more rows
#> # ℹ 27 more variables: stays_in_weekend_nights <dbl>,
#> #   stays_in_week_nights <dbl>, adults <dbl>, children <dbl>,
#> #   babies <dbl>, meal <fct>, country <fct>, market_segment <fct>,
#> #   distribution_channel <fct>, is_repeated_guest <dbl>,
#> #   previous_cancellations <dbl>, previous_bookings_not_canceled <dbl>,
#> #   reserved_room_type <fct>, assigned_room_type <fct>, …
# }